+++ openedx_events/event_bus/avro/tests/test_schema.py
@@ -1,6 +1,7 @@
 """
 Tests for event_bus.avro.schema module
 """
+from typing import Dict, List
 from unittest import TestCase
 
 from openedx_events.event_bus.avro.schema import schema_from_signal
@@ -243,3 +244,26 @@ class TestSchemaGeneration(TestCase):
 
         with self.assertRaises(Exception):
             schema_from_signal(DICT_SIGNAL)
+
+    def test_throw_exception_to_list_or_dict_types_without_annotation(self):
+        LIST_SIGNAL = create_simple_signal({"list_input": List})
+        DICT_SIGNAL = create_simple_signal({"list_input": Dict})
+        with self.assertRaises(TypeError):
+            schema_from_signal(LIST_SIGNAL)
+
+        with self.assertRaises(TypeError):
+            schema_from_signal(DICT_SIGNAL)
+
+    def test_list_with_annotation_works(self):
+        LIST_SIGNAL = create_simple_signal({"list_input": List[int]})
+        expected_dict = {
+            'name': 'CloudEvent',
+            'type': 'record',
+            'doc': 'Avro Event Format for CloudEvents created with openedx_events/schema',
+            'fields': [{
+                'name': 'list_input',
+                'type': {'type': 'array', 'items': 'long'},
+            }],
+        }
+        schema = schema_from_signal(LIST_SIGNAL)
+        self.assertDictEqual(schema, expected_dict)